Moved the signup form into it's own file
